Decouple the history and l2stateprovider packages
These two packages mostly coordinate through a History struct which is
actually more like a protocol-level concern. By moving the definition
of what can be in a history commitment to the chain-abstraction
package we can lower the coupling between the two packages.

This change also reenables the prefix proof tests which were commented
out because it was difficult not to have circular dependencies before
the refactoring. Ultimately, the test is moved to a new
testing/integration package to reflect that it tests the integration
among several other packages.