Simplify how hashing is done
Warning: The code will not panic or error if for some reason the
crypto.KeccackState Write or Read methods error out. However, this is
the same behavior of calling the crypto.Keccak256Hash().

This change also formats the comments to all be fewer than 80 columns
and removes some comments that were just describing what the code
does.