Further tighten up the hash functions
The main significance is making the hash and hashInto functions
operates on pointers to common.Hash instances instead of byte slices.
This is better because it limits the size of data for each pointer to
32 bytes (since `common.Hash` is a type alias for `[32]byte`).

This strong typing of the arguments means it is fine for the code to
ignore possible errors from the KeccakState Write and Read methods.