Isolate the legacy history commitment implementation
The legacy in inclusion-proofs packages are only used from tests at
this point. And, those tests are largely only used for benchmarking
the optimized implementation against the legacy one.

There is one method from the prefix-proofs package which is used in
the layer2-state-provider for checking that the prefix proofs
generated from the optimized implementation are verifiable.