Refactor, comment, and introduce a cursor
This change is mostly just suffling around code blocks, renaming
variables, and adding comments. But, it also adds a cursor and keeps
track of which node in the tree is actively being calculated.

This allows the historyCommitter to search for the values of specific
nodes by position in the tree while calculating the root.

The only reason for committing the code at this state is to benchmark
before and after adding the inclusion_proof calculation.