Add the lastLeafProver and wire it into root computation
This allows us to remove the whole implementation from
inclusion_proof.go and the associated tests. Also, fun fact, that
existing implementation was not only much slower, but it was
incorrect.

I'll add some differential fuzzing to show that these inclusion proofs
match the results from the legacy inclusionproofs package.