Remove the completeRoot method
This version of the code should be easier to maintain and only
performs minutely slower because of having to copy the cursor in each
recursive stack frame.

This commit also rearranges the conditional logic on the left and
right sides of the tree to make it simpler to follow the recursion.