Switch back to using the `InboxMessageCount` variable from previous
The problem with using the claimed assertion's `AfterState.Batch` is
that, in the case of assertions which overflow the maximum number of
blocks for an assertion, the `AfterState.Batch` will be lower than the
actual upper limit on batches. And, if a history commitment is
calculated based on that lower limit, it might not include all of the
blocks to which the assertion should commit.