Add FromHeight back to the HistoryCommitmentRequest
The problem with removing it is that the history commitments don't
always commit to leaves starting from the first leaf, `FromHeight=0`.

When there is a bisection, of a challenge edge, the history commitment
of one edge should be `FromHeight=0` `UpToHeight=(parentHeight/2)+1`
and the second edge should be `FromHeight=(parentHeight/2)+1` and
`UpToHeight=parentHeight`.