Decopule the callenge manager and assertion manager
Before this change, there were some flaky tests in the assertion manager
package. In part, this was because it was difficult to reason about the
responsibilities of each component.

Some of the work that the assertion manager was doing by making calls to its
logical parent (the challenge maanger) has been moved up to the challenge
manager with a very narrow interface remaning between the two.

Also, the construction of the challenge manager and its dependencies has been
slightly decoupled. Now the challenge manager takes in an almosst completetely
constructed assertion manager and registers itself as the RivalHandler for the
assertion manager. It also sets the correct address for the assertion manager
which is still being fetched from an on-chain contract during the construction
of the challenge manager. But, this can also be pulled out in a future commit
and passed into both objects (as well as the watcher and tracker objects.)

Another small code-quality improvement was pulling out the configuration that
affects the timing of various things the assertion manager is responsible for
into a separate private `times` struct helps to document the behavior of
those timers and also makes it easier to pass consistent values when testing.

Next up, I'd really like to pull some of the other types which are currently
being constructed during the construction of a challenge manager out and
improve on our ability to share the wiring across objects.