Move LayerZeroHeights and ChallengePeriodBlocks to chain abstraction
These values live on-chain, and should be read and cached during the
intialization of the chain abstraction to make them trival to retrieve in the
places where they are needed.

This commit also changes the type of the members of the LayerZeroHeights to a
protocol.Height and transitions to returning a LayerZeroHeights value instead of
a pointer to keep the shared pointer from being available to code that might
accidentally mutate it.