Allow a HeaderProvider to be passed to the stack
The desire is to be able to be responsive to new blocks as they are created on
the parent chain. There was some concern about relying on websockets for this
feed of headers, but nitro already has an inprocess abstraction which provides a
stream of parent-chain block headers.

The nitro codebase already has an implementation of the HeaderProvider interface
that can be passed into the challenge stack, and a tivial one has been added for
the end2end testing. Perhaps, after we merge the repos, we can just use the real
implementation from the nitro codebase in the e2e_test.go

Part of NIT-2793